debundle/lowering: extract body_facts (ModuleBodyFacts, RefCollector)
BAZEL_TEST_INVOCATIONS=none: file move.
