+++ devinfra/js/debundle/lowering/mod.rs
@@ -32,6 +32,7 @@ use spec::{
 };
 
 mod anonymous;
+mod body_facts;
 mod chunk_ast;
 mod chunk_renames;
 mod naturalize;
@@ -40,6 +41,7 @@ mod runtime_imports;
 mod visitors;
 
 use anonymous::resolve_anonymous_statement_ordinals;
+use body_facts::{ModuleBodyFacts, RefCollector, collect_module_body_facts};
 use chunk_ast::{
     ChunkAstAnalysis, TopLevelDecl, analyze_chunk_ast, binding_names, declaration_names,
     top_level_declaration_ids,
@@ -1218,22 +1220,6 @@ struct LowerChunkInputs<'a> {
     pre_existing_entry_exports: &'a BTreeSet<String>,
 }
 
-#[derive(Debug, Default)]
-struct ModuleBodyFacts {
-    /// Local bindings declared by an `import` statement in this module body.
-    /// Keyed by the binding's hygiene-aware `Id` so two same-named bindings
-    /// from different scopes are distinct.
-    imported_locals: HashSet<Id>,
-    /// All locals this module body provides — `imported_locals` plus
-    /// top-level declarations. Keyed by `Id`.
-    provided_locals: HashSet<Id>,
-    /// Every identifier the module body references (read or write).
-    /// Post-naturalize: any binding the heuristic naturalizer renamed
-    /// appears here under the post-rename `(sym, ctxt)` (the rename
-    /// mutates `sym` in place; `ctxt` is preserved).
-    referenced_idents: HashSet<Id>,
-}
-
 #[derive(Debug)]
 struct ImportedReexport {
     local: String,
@@ -2193,85 +2179,6 @@ fn known_effect_from_member_effect(effect: MemberEffect) -> Option<KnownEffect> 
     }
 }
 
-#[derive(Default)]
-struct RefCollector {
-    /// `Id`-keyed references collected from the body. The hygiene-aware
-    /// `Id = (sym, ctxt)` distinguishes same-named bindings declared in
-    /// different scopes (a function parameter `x` vs. a module-level `x`).
-    ids: HashSet<Id>,
-    /// Shadowing tracked by `sym` — kept for backwards-compatible behavior
-    /// with the pre-hygiene collector. With hygiene-correct contexts the
-    /// shadowed-by-sym filter is mostly redundant (the inner-scope ident
-    /// has its own `ctxt`, distinct from any outer reference), but the
-    /// filter preserves the exact set of `referenced_idents` produced
-    /// before the `Id` migration so downstream comparisons (e.g. against
-    /// String-keyed `declaration_by_name` via `sym`) don't drift.
-    shadowed_scopes: Vec<BTreeSet<String>>,
-}
-
-impl Visit for RefCollector {
-    fn visit_ident(&mut self, node: &Ident) {
-        let name = node.sym.as_ref();
-        if !self.is_shadowed(name) {
-            self.ids.insert(node.to_id());
-        }
-    }
-
-    fn visit_binding_ident(&mut self, _node: &BindingIdent) {}
-
-    fn visit_import_decl(&mut self, _node: &ImportDecl) {}
-
-    fn visit_function(&mut self, node: &Function) {
-        let shadowed = node
-            .params
-            .iter()
-            .flat_map(|param| binding_names(&param.pat))
-            .collect::<BTreeSet<_>>();
-        self.with_shadowed_scope(shadowed, |collector| node.visit_children_with(collector));
-    }
-
-    fn visit_arrow_expr(&mut self, node: &ArrowExpr) {
-        let shadowed = node
-            .params
-            .iter()
-            .flat_map(binding_names)
-            .collect::<BTreeSet<_>>();
-        self.with_shadowed_scope(shadowed, |collector| node.visit_children_with(collector));
-    }
-
-    fn visit_member_expr(&mut self, node: &MemberExpr) {
-        node.obj.visit_with(self);
-        if let MemberProp::Computed(computed) = &node.prop {
-            computed.expr.visit_with(self);
-        }
-    }
-
-    fn visit_prop_name(&mut self, node: &PropName) {
-        if let PropName::Computed(computed) = node {
-            computed.expr.visit_with(self);
-        }
-    }
-
-    fn visit_jsx_element_name(&mut self, _node: &JSXElementName) {}
-
-    fn visit_jsx_attr_name(&mut self, _node: &JSXAttrName) {}
-}
-
-impl RefCollector {
-    fn is_shadowed(&self, name: &str) -> bool {
-        self.shadowed_scopes
-            .iter()
-            .rev()
-            .any(|scope| scope.contains(name))
-    }
-
-    fn with_shadowed_scope<F: FnOnce(&mut Self)>(&mut self, names: BTreeSet<String>, f: F) {
-        self.shadowed_scopes.push(names);
-        f(self);
-        self.shadowed_scopes.pop();
-    }
-}
-
 /// Drop `ImportSpecifier::Named` specifiers from a residual entry
 /// body whose locals are unused after a logical-module move and
 /// whose binding name is claimed by `Schedule.bindings`. If all
@@ -2368,30 +2275,6 @@ fn reject_duplicate_member_bindings(
         );
     }
     Ok(())
-}
-
-fn collect_module_body_facts(body: &[ModuleItem]) -> ModuleBodyFacts {
-    let mut facts = ModuleBodyFacts::default();
-    let mut ref_collector = RefCollector::default();
-    for item in body {
-        item.visit_with(&mut ref_collector);
-        facts
-            .provided_locals
-            .extend(top_level_declaration_ids(item));
-        if let ModuleItem::ModuleDecl(ModuleDecl::Import(import)) = item {
-            for specifier in &import.specifiers {
-                let local = match specifier {
-                    ImportSpecifier::Named(named) => named.local.to_id(),
-                    ImportSpecifier::Default(default) => default.local.to_id(),
-                    ImportSpecifier::Namespace(namespace) => namespace.local.to_id(),
-                };
-                facts.imported_locals.insert(local.clone());
-                facts.provided_locals.insert(local);
-            }
-        }
-    }
-    facts.referenced_idents = ref_collector.ids;
-    facts
 }
 
 fn collect_imported_reexports_by_module(
